salesforce: move toUTF8 out to utils
