+++ packages/salesforce/src/Utils.js
@@ -51,6 +51,32 @@ function createAccessTokenConnection(state) {
   };
 }
 
+let anyAscii = undefined;
+
+// use a dynamic import because any-ascii is pure ESM and doesn't play well with CJS
+// This promise MUST be resolved by execute before a connection is created
+export const loadAnyAscii = state =>
+  import('any-ascii').then(m => {
+    anyAscii = m.default;
+    return state;
+  });
+
+/**
+ * Transliterates unicode characters to their best ASCII representation
+ * @public
+ * @example <caption>Transliterate `άνθρωποι` to `anthropoi`</caption>
+ * fn((state) => {
+ *   const s = toUTF8("άνθρωποι");
+ *   console.log(s); // anthropoi
+ *   return state;
+ * });
+ * @param {string} input - A string with unicode characters
+ * @returns {string} - ASCII representation of input string
+ */
+export function toUTF8(input) {
+  return anyAscii(input);
+}
+
 /**
  * Creates a connection to Salesforce using Basic Auth or OAuth.
  * @function createConnection
